hastighet på bullets när MAN springer

+++ Tempo/src/tempo/sprites/Player.java
@@ -11,9 +11,9 @@ import java.awt.image.*;
 
 public class Player extends Sprites {
     
-    public double xSpeed = 0.0; 
-    public double ySpeed = 0.0;
-    public double gravity = 0;
+    public float xSpeed = 0; 
+    public float ySpeed = 0;
+    public float gravity = 0;
     public int jumps = 0;
     public int nuts = 5;
     public int maxNuts = 5;
@@ -45,6 +45,7 @@ public class Player extends Sprites {
         );
                     }
     }
-    
+        
+       
     
 }
